My story
👋 Hi, I’m @Danilato
👀 I’m interested in 3d modeling and simulation.

Resently I recently created a short film. This is completely CG animation. https://vimeo.com/583343965 

I was interested in the simulation of liquids and their variability depending on my control and not on the control of the computer. What percentage of the simulation is really controlled by us, that's the question. The process of creating and searching can be viewed here https://vimeo.com/583363016v

🌱 I’m currently learning Python with 3d-modeling. Also (may be it sounds little bit strange...) Python in Crypto and 3d-modelling in Crypto.

💞️ I’m looking to collaborate on ...a lot of projects and a lot of ideas in my mind. And it will be disrespect for all to tell someone name.

📫 How to reach me danilakornev@yandex.ru